Migrate SelftestDataCreate to Xoshiro256ss in Utils.pas
The reason is that FPC devs are planning to change the default RNG
(see https://wiki.freepascal.org/User_Changes_Trunk#Random_generator for
details), and we need SelftestDataCreate to behave deterministically, so port
it to the generator that never changes.

However, this commit breaks backward compatibility and generates different data
compared to the previous one.

Also, the number of games is extended to 128 from 100.